add instance norm and fixed software emit verilog
