+++ machop/mase_components/norm/test/group_norm_2d_tb.py
@@ -15,7 +15,7 @@ from cocotb.result import TestFailure
 from cocotb.triggers import *
 
 from mase_cocotb.testbench import Testbench
-from mase_cocotb.interfaces.streaming import StreamDriver, StreamMonitor
+from mase_cocotb.interfaces.streaming import StreamDriver, ErrorThresholdStreamMonitor
 from mase_cocotb.runner import mase_runner
 from mase_cocotb.matrix_tools import (
     gen_random_matrix_input,
@@ -26,13 +26,10 @@ from mase_cocotb.utils import (
     bit_driver,
     batched,
     sign_extend_t,
-    sign_extend,
-    random_2d_dimensions,
 )
 
 from chop.passes.graph.transforms.quantize.quantized_modules import (
     GroupNormInteger,
-    LayerNormInteger
 )
 from chop.passes.graph.transforms.quantize.quantizers.quantizers_for_hw import (
     integer_floor_quantizer_for_hw,
@@ -160,67 +157,6 @@ def _fixed_group_norm_2d_model(
         "diff": diff_int,
         "norm": norm_int,
     }
-
-
-class ErrorThresholdStreamMonitor(StreamMonitor):
-    def __init__(
-        self,
-        clk,
-        data,
-        valid,
-        ready,
-        width: int,       # Width of the number
-        signed: bool,     # Signedness of number
-        error_bits: int,  # Number of last bits the number can be off by
-        check=True,
-        name=None,
-    ):
-        super().__init__(clk, data, valid, ready, check, name)
-
-        self.width = width
-        self.signed = signed
-        self.error_bits = error_bits
-        self.log.setLevel("INFO")
-
-    def _check(self, got, exp):
-        fail = not self.check
-        if type(got) != type(exp):
-            assert fail, (
-                f"Type Mismatch got:{type(got)} vs. exp:{type(exp)}"
-            )
-
-        # Compare Outputs
-        if type(got) == list:
-            g = np.array(got)
-            e = np.array(exp)
-            if self.signed:
-                g = sign_extend(g, self.width)
-                e = sign_extend(e, self.width)
-            err = np.abs(g - e)
-            max_biterr = np.full_like(err, self.error_bits)
-            if not (err <= max_biterr).all():
-                self.log.error(
-                    "Failed | Got: %20s Exp: %20s Err: %14s" % (g, e, err)
-                )
-                assert fail, "Test Failed!"
-                return
-
-        elif type(got) == int:
-            g, e = got, exp
-            if self.signed:
-                g = sign_extend(g, self.width)
-                e = sign_extend(e, self.width)
-            err = abs(g - e)
-            if not err <= self.error_bits:
-                self.log.error(
-                    "Failed | Got: %20s Exp: %20s Err: %10s" % (g, e, err)
-                )
-                assert fail, "Test Failed!"
-                return
-
-        self.log.debug(
-            "Passed | Got: %20s Exp: %20s Err: %10s" % (g, e, err)
-        )
 
 
 class GroupNorm2dTB(Testbench):
